PythonSK/requirements.txt
semantic-kernel==1.41.0
streamlit==1.32.0
python-dotenv==1.0.1
pyyaml==6.0.1

